Refactored home - split home into pages (on their own js files)
