+++ src/components/Navigation/Toolbar/Toolbar.css
@@ -62,7 +62,7 @@
 
 #toolbar-navbar {
   background-color: transparent;
-  padding: 7em 7em 7em 9em;
+  padding: 5em 5em 5em 7em;
 }
 
 .ui.left.visible.sidebar {
